+++ src/main/java/com/omar/model/game/README.md
@@ -1,0 +1,1 @@
+The code in HexGame and HexMech was copied and refactored from: https://gist.github.com/salamander2/4329783
