+++ src/main/java/com/omar/model/upperpanel/MainPanel.java
@@ -1,4 +1,4 @@
-package com.omar.view.upperpanel;
+package com.omar.model.upperpanel;
 
 import javax.swing.*;
 import javax.swing.border.LineBorder;
