--- src/main/java/com/omar/view/hexpanel/README.md
@@ -1,1 +1,0 @@
-This code was copied and refactored from: https://gist.github.com/salamander2/4329783
