--- src/test/java/com/omar/model/faction/FactionTest.java
@@ -1,20 +1,0 @@
-package com.omar.model.faction;
-
-import org.junit.jupiter.api.Test;
-
-import java.util.LinkedList;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class FactionTest {
-    Faction sampleFaction;
-    public FactionTest(){
-        this.sampleFaction = new Faction("Redosia", 15, new LinkedList<>());
-    }
-    @Test
-    void testInitArmies(){
-        assertEquals("Redosia", sampleFaction.getName());
-        assertEquals(15, sampleFaction.getArmy(0).getPosition());
-        assertEquals(10, sampleFaction.getArmy(0).getFirepower());
-    }
-}
